etc: Support local customizations in *.inc
This is useful for places, like hashbang.sh, which have site-specific
modifications of the *.inc files.

With the current setup, the package manager cannot automatically install
updated versions of those files, as it would need to somehow merge the
site-specific and upstream changes.  Having the site-specific changes in
separate files solves this.