qbittorrent: add netlink protocol, private-etc
Netlink protocol is needed if user select to bind specific network interface in config. Otherwise it throws an error in qbittorent log:
The network interface defined is invalid: tun0

Example private-etc is added but commented out by default. It's tested but as there are many different system configurations users should enable it manually.