Add Debian builds to .travis.yml (work in progress).
This experimentally adds Debian package building for our Travis
CI service. This is needed because there are errors building
Debian packages even when firejail as a whole builds correctly
due to errors in firejail's Debian conffiles. For instance, at
the moment `make deb` fails because lxterminal is not listed in
platform/debian/conffiles