blacklist ksslcertificatemanager
While it is believed that blacklisting these files is a safe default,
it has the effect that untrusted certificates have to be acknowledged every
time they are encountered (with whitelisting it is possible to accept
them for the duration of an application session).

Where this causes usability issues, it will be necessary to noblacklist
these paths.