Revert "apparmor fix: somehow it cannot find the firejail profile to load it"
This reverts commit 949a221a1b92e422e6dcb7ea6089ed5c8d5cc22a.

The 'firejail-default' is the name of 'unnatached' profile not path
to it. Moreover names starting with '/' are changing profile type
back to 'standard' which in this case means we literally create
profile for the profile file itself '/etc/apparmor.d/firejail-default'.
That means firejail would never load this profile to contain any
app thus we have to revert this. For more info, see
https://www.suse.com/documentation/sles-15/singlehtml/book_security/book_security.html#sec.apparmor.profiles.types.unattached