Begin using gitlab for building to save time on auto code tests.
Code tests are still run on Travis CI, but distro-specific builds are run on GitLab CI.
Currently these are used:
1. Debian stable
2. Ubuntu latest
3. CentOS latest
4. Fedora latest
5. Alpine

Debian and CentOS are for testing builds on older systems, for *.deb vs *.rpm, respectively.
Ubuntu and Fedora are for testing builds on the latest **stable** systems for *.deb and *.rpm.
Alpine is used to test building/installing from source.

All run concurrently.

In the future may expand tests on Gitlab to cover code testing as well.