Allow Tor Browser to run /usr/bin/id (#3114)
The start-tor-browser script tries to run /usr/bin/id to check that
it isn't root before starting the browser. See
https://gitweb.torproject.org/builders/tor-browser-build.git/tree/projects/tor-browser/RelativeLink/start-tor-browser?id=41fd236bbb7d3d75a27473f927be31f7dd8fdc99#n94

If id is not in the private-bin directory, the test still works by
accident, but prints these error messages:

  ./Browser/start-tor-browser: line 94: id: command not found
  ./Browser/start-tor-browser: line 94: [: : integer expression expected

Add id to the private-bin directory to make it run as intended.