steam.profile: correctly blacklist unneeded directories in user's home
"noblacklist" directives prevent following ones from blacklisting the specified directory/file.

The profile currently has a "noblacklist" directive for each directory used by Steam and/or its games, which is fine.

However, there are no directives blacklisting the user's home, thus all directories and files inside it are accessible by Steam.

This commit fixes the issue by adding "whitelist" directives, which automatically blacklist the parent directory (in this case the user's home).

"mkdir" and "mkfile" directives are added so that the directories/files are created if they don't exist.

Thanks to @SkewedZeppelin for suggesting to keep "noblacklist" and use "mkdir" and "mkfile".