manpage: /bin/bash -> user's perferred shell
We do not start /bin/bash in the sandbox, we use $SHELL (which is
usually /bin/bash). See #3434 and #3844. This commit updates the
manpage accordingly until #3434 is resolved with a final solution like
using /bin/bash or /bin/sh as hardcoded default. Close #3844.

The descriptions of --join* are not updated as there is currenly some
work, see #2934 and #3850.