signal-desktop.profile: fix typo of disable-xdg.profile
Added on commit f4f676745 ("Refactor electron.profile and electron based
programs (#3807)").

This appears to be the only instance of that:

    $ grep -Fnr 'include-xdg' etc
    etc/profile-m-z/signal-desktop.profile:9:ignore include-xdg.inc

Simply fixing the typo would enable xdg dirs for the first time since
the aforementioned commit.  But, as talked with @rusty-snake[1], since
there has been no negative feedback, and since it's a whitelisting
profile, just remove the affected line instead.

Credits go to syntax highlighting on vim.

[1]: https://github.com/netblue30/firejail/pull/4001