man: corrections regarding --private-FOO options
Commit 0.9.60-1070-g40d3604f updated the man pages with respect to
--private-opt, --private-etc, and --private-srv.  It was made after
testing firejail 0.9.52 (from Ubuntu 18.04).  However, it
unfortunately did not accurately reflect the the behavior of the
current HEAD at the time, because commit 0.9.56-rc1-14-ga9242301 had
previously slightly changed the behavior of these three options (after
0.9.52), and was released in 0.9.56.  The man pages changes made in
commit 40d3604f were therefore not entirely correct.

This commit updates the man pages to describe the behavior as
implemented in a9242301 (and is still the behavior as of the current
HEAD: 0.9.64-737-g937815ba).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>