Fix #4282 -- Unable to open X display when running firejail chromium command
Summary: SDDM uses $XDG_RUNTIME_DIR/<UUID> as Xauthority.

In my tests (Fedora 32 KDE spin IIRC) it used /tmp/... so it was
irrelevant for wruc. So the Xauthority file created by SDDM sems to
depend on distro, version, config, ….

Future alternatives to this long, ugly line would be a ${XAUTHORITY}
macro or a private-run-user option.