cmdline.c: optionally quote the resulting command line
If we were launched by sshd, do not add extra quotes to the command
line.  This is because if firejail is a login shell, sshd will launch
firejail thusly:

* argv[0]: /path/to/firejail
* argv[1]: -c
* argv[2]: user's command to execute

For example, if the user executed "ssh othernode echo hello world",
argv[2] will be "echo hello world".  Firejail will then add *extra*
quotes to it, resulting in argv[2] becoming "'echo hello world' "
(without the "", of course).  The user's shell (e.g., bash) will see
the extra single quotes and will not split the token into multiple
tokens.  The shell will be unable to find an executable or intrinsic
named "echo hello world ", so it will fail.

This commit changes the above behavior if firejail is launched by
sshd.  In that case, firejail will *not* add the extra single quotes
around argv[2].  Specifically: all the tokens still end up in argv[2],
but there's no *extra* quotes around argv[2], so the shell will split
argv[2] into multiple tokens (if necessary).  In the above example,
argv[2] will be "echo hello world" (without the ""), which will be
split.  The shell will then look for an intrinsic or executable named
"echo", which will succeed, and "hello world" will ultimately be
emitted.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>