Revert "allow/deny help and man pages"
This reverts commit a11707ea273e5665047f8a7d9387ba07f08d72f6.

The man pages currently direct users to use the aliases instead of the
commands, which some users of firejail-git may end up doing.  Example:

https://github.com/netblue30/firejail/discussions/4496

So revert the man page changes as well to avoid confusion.

Note: This is not a full revert.  The commit in question also contains
some string formatting fixes on src/firejail/usage.c (related to dbus
and netmask), which are left intact.

Relates to #4410.