Remove profcleaner.c and profcleaner.sh
As of this commit, these are not of much use.  Though later if a generic
profile search/replace tool with built-in rules is to be added, the
tools in question could be used as a starting point.

src/tools/profcleaner.c was added on commit fe0f975f4 ("move
whitelist/blacklist to allow/deny", 2021-07-05).

src/tools/profcleaner.sh was added on commit ed02ab57b ("Create
profcleaner.sh", 2021-07-07) / PR #4389.

Relates to #4410.