RELNOTES: add bugfixes
Note: They are added in the order that the issues were fixed/closed.

Note2: The issues were found through the following url:

https://github.com/netblue30/firejail/issues?q=is%3Aclosed+label%3Abug+-label%3Asecurity+closed%3A%3E2021-06-29+

The date used is the release date of 0.9.66, so in theory the query
should return every bug closed after that.  Security-related issues are
excluded because they will be added separately.

Note3: All issues other than #4328 were fixed before 0.9.68rc1.

Relates to #2758 #4235 #4328 #4387 #4395 #4460 #4467 #4558 #4560 #4586.