testing: fix expect matching of numbers
The sandbox timestamp should not be available for matching
when output is already expected from the next command
(this is only a problem if numeric output if expected from the
first command in the sandbox).

A possible alternative would be to flush the expect output buffer
with 'expect "*"' after the sandbox is up.