gcov: fix gcov functions always declared as dummy
Currently, the check to enable gcov relies on a non-existent macro due
to a typo, which looks like it would cause the dummy/empty versions of
the gcov functions to always be declared (even with --enable-gcov),
instead of the real ones from gcov.h.  This commit fixes the typo
(HAS_GCOV -> HAVE_GCOV).  See configure.ac for the macro declaration.

This amends commit 5106b2ec4 ("gcov: use no-op functions if not
enabled", 2021-06-20) / PR #4376.

Occurrences of each macro with this commit applied:

    $ git grep -F HAVE_GCOV | wc -l
    16
    $ git grep -F HAS_GCOV | wc -l
    0