makefiles: equalize object dependencies in program targets
Compared to the objects that are actually used in a given recipe, some
program targets are missing object dependencies, while others appear to
have unused object dependencies.

Make each of those targets depend on the objects that are actually used
when linking.

Note: No check was done for extraneous/missing objects when linking;
this commit only makes the object dependencies equal to the objects
that are linked.